+++ test/object-store/sync/flx_sync.cpp
@@ -287,6 +287,83 @@ TEST_CASE("app: error handling integration test", "[sync][flx][baas]") {
     config.sync_config->error_handler = std::move(error_handler);
     config.sync_config->client_resync_mode = ClientResyncMode::Manual;
 
+    SECTION("Resuming while waiting for session to auto-resume") {
+        enum class TestState { InitialSuspend, InitialResume, SecondBind, SecondSuspend, SecondResume, Done };
+        TestingStateMachine<TestState> state(TestState::InitialSuspend);
+        config.sync_config->on_sync_client_event_hook = [&](std::weak_ptr<SyncSession>,
+                                                            const SyncClientHookData& data) {
+            std::optional<TestState> wait_for;
+            auto event = data.event;
+            state.transition_with([&](TestState state) -> std::optional<TestState> {
+                if (state == TestState::InitialSuspend && event == SyncClientHookEvent::SessionSuspended) {
+                    // If we're getting suspended for the first time, notify the test thread that we're
+                    // ready to be resumed.
+                    wait_for = TestState::SecondBind;
+                    return TestState::InitialResume;
+                }
+                else if (state == TestState::SecondBind && data.event == SyncClientHookEvent::BindMessageSent) {
+                    return TestState::SecondSuspend;
+                }
+                else if (state == TestState::SecondSuspend && event == SyncClientHookEvent::SessionSuspended) {
+                    wait_for = TestState::Done;
+                    return TestState::SecondResume;
+                }
+                return std::nullopt;
+            });
+            if (wait_for) {
+                state.wait_for(*wait_for);
+            }
+            return SyncClientHookAction::NoAction;
+        };
+        auto r = Realm::get_shared_realm(config);
+        wait_for_upload(*r);
+        nlohmann::json error_body = {
+            {"tryAgain", true},           {"message", "fake error"},
+            {"shouldClientReset", false}, {"isRecoveryModeDisabled", false},
+            {"action", "Transient"},      {"backoffIntervalSec", 900},
+            {"backoffMaxDelaySec", 900},  {"backoffMultiplier", 1},
+        };
+        nlohmann::json test_command = {{"command", "ECHO_ERROR"},
+                                       {"args", nlohmann::json{{"errorCode", 229}, {"errorBody", error_body}}}};
+
+        // First we trigger a retryable transient error that should cause the client to try to resume the
+        // session in 5 minutes.
+        auto test_cmd_res =
+            wait_for_future(SyncSession::OnlyForTesting::send_test_command(*r->sync_session(), test_command.dump()))
+                .get();
+        REQUIRE(test_cmd_res == "{}");
+
+        // Wait for the
+        state.wait_for(TestState::InitialResume);
+
+        // Once we're suspended, immediately tell the sync client to resume the session. This should cancel the
+        // timer that would have auto-resumed the session.
+        r->sync_session()->handle_reconnect();
+        state.transition_with([&](TestState cur_state) {
+            REQUIRE(cur_state == TestState::InitialResume);
+            return TestState::SecondBind;
+        });
+        state.wait_for(TestState::SecondSuspend);
+
+        // Once we're connected again trigger another retryable transient error. Before RCORE-1770 the timer
+        // to auto-resume the session would have still been active here and we would crash when trying to start
+        // a second timer to auto-resume after this error.
+        test_cmd_res =
+            wait_for_future(SyncSession::OnlyForTesting::send_test_command(*r->sync_session(), test_command.dump()))
+                .get();
+        REQUIRE(test_cmd_res == "{}");
+        state.wait_for(TestState::SecondResume);
+
+        // Finally resume the session again which should cancel the second timer and the session should auto-resume
+        // normally without crashing.
+        r->sync_session()->handle_reconnect();
+        state.transition_with([&](TestState cur_state) {
+            REQUIRE(cur_state == TestState::SecondResume);
+            return TestState::Done;
+        });
+        wait_for_download(*r);
+    }
+
     SECTION("handles unknown errors gracefully") {
         auto r = Realm::get_shared_realm(config);
         wait_for_download(*r);
@@ -401,6 +478,7 @@ TEST_CASE("app: error handling integration test", "[sync][flx][baas]") {
         REQUIRE(error.is_client_reset_requested());
         REQUIRE(error.is_fatal);
     }
+
 
     SECTION("teardown") {
         harness.reset();